+++ src/Styles/App.css
@@ -14,25 +14,22 @@
   pointer-events: none;
 }
 
+.custom-loading-text {
+  text-align: center;
+}
+
+.custom-loading-text.show {
+  display: block;
+}
+
+.custom-loading-text.hide {
+  display: none;
+}
+
 @media (prefers-reduced-motion: no-preference) {
   .App-logo {
     animation: App-logo-spin infinite 20s linear;
   }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
 }
 
 @keyframes App-logo-spin {
