+++ src/Styles/Logo.css
@@ -1,6 +1,6 @@
 .custom-loading-text {
-    color: #ffffff;
-    font-size: 24px; 
-    font-family: 'Lobster', cursive;
+  color: #ffffff;
+  font-size: 24px; 
+  font-family: 'Lobster', cursive;
 }
